+++ src/main/java/com/codeit/hrbank/domain/backup/dto/request/CursorBackupRequestDto.java
@@ -11,6 +11,6 @@ public record CursorBackupRequestDto(
         LocalDateTime startedAtTo,
         String sortField,
         String sortDirection,
-        Long size
+        int size
 ) {
 }
